Create in_theater.rb
Exercise in Launch Academy week 1

method get_highest_rating is re-written in a better way to access necessary information.

get_cast_members and above is not really optimize.

To test this method, file "in_theaters.json" will be needed. You can find it in the same folder as this in_theater.rb file.